requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
openai
jinja2
python-dotenv
